+++ packages/orq-rc/src/orq_ai_sdk/models/createbudgetop.py
@@ -241,5 +241,5 @@ class CreateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-27T14:02:12.190Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-28T05:14:50.077Z")
     r"""The date and time the resource was last updated"""