+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KB2SYKB0JQ3APTH8R8CAP8HA"
+        "01KB4E5QNP0703RFEKJ0EAFZDG"
     )
     r"""The unique identifier of the data source"""