+++ packages/orq-rc/src/orq_ai_sdk/models/parseop.py
@@ -24,7 +24,7 @@ class AgenticChunkerStrategyTypedDict(TypedDict):
     r"""The text content to be chunked"""
     strategy: AgenticChunker
     model: str
-    r"""Chat model to use for chunking. (Available models)[https://docs.orq.ai/docs/proxy#chat-models]"""
+    r"""Model to use for chunking. (Available models)[https://docs.orq.ai/docs/proxy/supported-models#chat-models]"""
     metadata: NotRequired[bool]
     r"""Whether to include metadata for each chunk"""
     return_type: NotRequired[ParseChunkingRequestChunkingRequestReturnType]
@@ -46,7 +46,7 @@ class AgenticChunkerStrategy(BaseModel):
     strategy: AgenticChunker
 
     model: str
-    r"""Chat model to use for chunking. (Available models)[https://docs.orq.ai/docs/proxy#chat-models]"""
+    r"""Model to use for chunking. (Available models)[https://docs.orq.ai/docs/proxy/supported-models#chat-models]"""
 
     metadata: Optional[bool] = True
     r"""Whether to include metadata for each chunk"""
@@ -99,7 +99,7 @@ class SemanticChunkerStrategyTypedDict(TypedDict):
     r"""The text content to be chunked"""
     strategy: SemanticChunker
     embedding_model: str
-    r"""Embedding model to use for semantic similarity. (Available embedding models)[https://docs.orq.ai/docs/proxy#embedding-models]"""
+    r"""Embedding model to use for semantic similarity. (Available embedding models)[https://docs.orq.ai/docs/proxy/supported-models#embedding-models]"""
     metadata: NotRequired[bool]
     r"""Whether to include metadata for each chunk"""
     return_type: NotRequired[ParseChunkingRequestChunkingReturnType]
@@ -127,7 +127,7 @@ class SemanticChunkerStrategy(BaseModel):
     strategy: SemanticChunker
 
     embedding_model: str
-    r"""Embedding model to use for semantic similarity. (Available embedding models)[https://docs.orq.ai/docs/proxy#embedding-models]"""
+    r"""Embedding model to use for semantic similarity. (Available embedding models)[https://docs.orq.ai/docs/proxy/supported-models#embedding-models]"""
 
     metadata: Optional[bool] = True
     r"""Whether to include metadata for each chunk"""