+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KB2SYKAYYZ4QX8HNGR07FPPG"
+        "01KB4E5QNN8BGKD2BCND5SWWGD"
     )
     r"""The unique identifier of the data source"""