+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KB2SYKB24NQD5KTJGG97JEYQ"
+        "01KB4E5QNRGTGMM8WA8RPMKT03"
     )
     r"""The unique identifier of the data source"""